refactor: extract shared deposit witness helpers
