+++ lib/src/deposit_circuit.rs
@@ -1,89 +1,10 @@
 //! This module contains the code for generating the deposit proofs, these are made by the proposer because they deal with
 //! sha256 hashes within a circuit.
-use crate::{nf_client_proof::PublicInputs, shared_entities::DepositData};
+use crate::{deposit_witness::DepositDataVar, nf_client_proof::PublicInputs, shared_entities::DepositData};
 use ark_bn254::Fr as Fr254;
 use ark_ff::{PrimeField, Zero};
-use jf_primitives::circuit::{poseidon::PoseidonHashGadget, sha256::Sha256HashGadget};
+use jf_primitives::circuit::sha256::Sha256HashGadget;
 use jf_relation::{errors::CircuitError, BoolVar, Circuit, PlonkCircuit, Variable};
-
-/// Storing the 'Variable's needed to crete a deposit commitment.
-/// Note that if it's a deposit_fee, nf_token_id and nf_slot_id will be fee_token_id,
-/// value will be deposit_fee. These are handled onchain automatically in escrow fund function.
-#[derive(Clone, Copy, Default, Debug)]
-pub struct DepositDataVar {
-    pub nf_token_id: Variable,
-    pub nf_slot_id: Variable,
-    pub value: Variable,
-    pub secret_hash: Variable,
-}
-
-impl DepositDataVar {
-    /// Creates the commitment hash from this [`DepositDataVar`].
-    pub fn to_commitment(
-        &self,
-        circuit: &mut PlonkCircuit<Fr254>,
-        flag: BoolVar,
-    ) -> Result<Variable, CircuitError> {
-        // Calculate the commitment.
-        let calculated_hash = circuit.poseidon_hash(&[
-            self.nf_token_id,
-            self.nf_slot_id,
-            self.value,
-            circuit.zero(),
-            circuit.one(),
-            self.secret_hash,
-        ])?;
-
-        // If flag is true we output the zero commitment, else the calculated commitment
-        circuit.conditional_select(flag, calculated_hash, circuit.zero())
-    }
-    /// Schedules the sha256 hash of this [`DepositDataVar`] to be calculated.
-    /// Returns the output of the hash right shifted by 4 so it is in agreement with Nightfall.sol.
-    pub fn sha256_and_shift(
-        &self,
-        circuit: &mut PlonkCircuit<Fr254>,
-        lookup_vars: &mut Vec<(Variable, Variable, Variable)>,
-        flag: BoolVar,
-    ) -> Result<Variable, CircuitError> {
-        let (_, sha256_var) = circuit.full_shifted_sha256_hash(
-            &[
-                self.nf_token_id,
-                self.nf_slot_id,
-                self.value,
-                self.secret_hash,
-            ],
-            lookup_vars,
-        )?;
-
-        // If flag is true then this was a dummy variable and so we want to output the zero hash.
-        circuit.conditional_select(flag, sha256_var, circuit.zero())
-    }
-    /// Returns a [`BoolVar`] indicating whether this real deposit data or dummied to be zero
-    pub fn is_real(&self, circuit: &mut PlonkCircuit<Fr254>) -> Result<BoolVar, CircuitError> {
-        // We check if both value and token id are zero, if this is the case then the this is a dummy entry and we output true.
-        let value_zero = circuit.is_zero(self.value)?;
-        let id_zero = circuit.is_zero(self.nf_token_id)?;
-        circuit.logic_and(value_zero, id_zero)
-    }
-
-    /// Creates a new instance of [`DepositDataVar`] from a [`DepositData`]
-    pub fn from_deposit_data(
-        data: &DepositData,
-        circuit: &mut PlonkCircuit<Fr254>,
-    ) -> Result<Self, CircuitError> {
-        let nf_token_id = circuit.create_variable(data.nf_token_id)?;
-        let nf_slot_id = circuit.create_variable(data.nf_slot_id)?;
-        let value = circuit.create_variable(data.value)?;
-        let secret_hash = circuit.create_variable(data.secret_hash)?;
-
-        Ok(DepositDataVar {
-            nf_token_id,
-            nf_slot_id,
-            value,
-            secret_hash,
-        })
-    }
-}
 
 pub trait DepositCircuitGadget<F>
 where
