+++ lib/src/lib.rs
@@ -6,6 +6,7 @@ pub mod commitments;
 pub mod constants;
 pub mod contract_conversions;
 pub mod deposit_circuit;
+pub mod deposit_witness;
 pub mod derive_key;
 pub mod error;
 pub mod health_check;
